refactor: 이름 구분자 형식 검증 및 split 기능 View로 이동
